+++ src/views/NavBar/index.tsx
@@ -1,34 +1,14 @@
 import React, { useCallback } from 'react'
-import { useAppDispatch, useAppSelector, thunkActions } from 'state'
 import { useHistory } from 'react-router-dom'
 import { useSignIn } from 'utils'
 import { SignedIn } from 'views/NavBar/SignedIn'
 import { NotSignedIn } from 'views/NavBar/NotSignedIn'
+import { AccountIcon } from 'views/NavBar/AccountIcon'
 
 import clsx from 'clsx'
 import { makeStyles, useTheme, Theme, createStyles } from '@material-ui/core/styles'
-import Drawer from '@material-ui/core/Drawer'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import AppBar from '@material-ui/core/AppBar'
-import Toolbar from '@material-ui/core/Toolbar'
-import List from '@material-ui/core/List'
-import Typography from '@material-ui/core/Typography'
-import Divider from '@material-ui/core/Divider'
-import IconButton from '@material-ui/core/IconButton'
-import MenuIcon from '@material-ui/icons/Menu'
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft'
-import ChevronRightIcon from '@material-ui/icons/ChevronRight'
-import ListItem from '@material-ui/core/ListItem'
-import ListItemText from '@material-ui/core/ListItemText'
-import MenuItem from '@material-ui/core/MenuItem'
-import Menu from '@material-ui/core/Menu'
-import HomeIcon from '@material-ui/icons/Home'
-import AccountCircle from '@material-ui/icons/AccountCircle'
-import ExitToAppIcon from '@material-ui/icons/ExitToApp'
-import HelpIcon from '@material-ui/icons/Help'
-import Avatar from '@material-ui/core/Avatar'
-import Fab from '@material-ui/core/Fab'
-import Add from '@material-ui/icons/Add'
+import { Drawer, CssBaseline, AppBar, Toolbar, List, Divider, IconButton, ListItem, ListItemText, Button } from '@material-ui/core'
+import { Menu, ChevronLeft, ChevronRight, Home, Help } from '@material-ui/icons'
 
 const drawerWidth = 240
 
@@ -101,12 +81,6 @@ const useStyles = makeStyles((theme: Theme) =>
 const NavBar: React.FC = () => {
   const signIn = useSignIn()
   const history = useHistory()
-  const dispatch = useAppDispatch()
-  const userInfo = useAppSelector(state => state.auth.userInfo)
-  let userImg = ''
-  if (userInfo.imgOwn === '1') {
-    userImg = `data:image/${userInfo.imgType}base64,${userInfo.img64}`
-  }
 
   const link = useCallback(
     (path: string) => () => {
@@ -116,8 +90,6 @@ const NavBar: React.FC = () => {
   const classes = useStyles()
   const theme = useTheme()
   const [draweropen, setOpen] = React.useState(false)
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
-  const useropen = Boolean(anchorEl)
 
   const handleDrawerOpen = () => {
     setOpen(true)
@@ -127,34 +99,11 @@ const NavBar: React.FC = () => {
     setOpen(false)
   }
 
-  const handleClose = () => {
-    setAnchorEl(null)
-  }
-
-  const handleMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget)
-  }
-
-  const linkClose = useCallback(
-    (path: string) => {
-      link(`/${path}`)()
-      setAnchorEl(null)
-    }, [setAnchorEl, link]
-  )
-
   const linkDrawerClose = useCallback(
     (path: string) => {
       link(`/${path}`)()
       setOpen(false)
     }, [setOpen, link]
-  )
-
-  const logoutClose = useCallback(
-    () => {
-        setAnchorEl(null)
-        dispatch(thunkActions.auth.signOut())
-        link('/')()
-      }, [dispatch, link]
   )
 
   return (
@@ -174,7 +123,7 @@ const NavBar: React.FC = () => {
             edge="start"
             className={clsx(classes.menuButton, draweropen && classes.hide)}
           >
-            <MenuIcon />
+            <Menu />
           </IconButton>
           <img
             src={`${process.env.PUBLIC_URL}/img/logo_transparent.png`}
@@ -187,76 +136,14 @@ const NavBar: React.FC = () => {
             alt='logo'
             onClick={link('/')}
           />
-          {signIn && (
-            <React.Fragment>
-              <Fab color='secondary' size='small' onClick={link('/games/new')}>
-                <Add/>
-              </Fab>
-              <Typography
-                variant="h6"
-                noWrap
-                style={{
-                  marginLeft: 'auto',
-                  fontSize: 18
-                }}
-              >
-                {userInfo.userName}
-              </Typography>
-              <IconButton
-                style={{
-                  //marginLeft: 'auto'
-                }}
-                aria-label="account of current user"
-                aria-controls="menu-appbar"
-                aria-haspopup="true"
-                onClick={handleMenu}
-                color="inherit"
-              >
-                <Avatar
-                  alt='userImage'
-                  src={userImg}
-                  style={{
-                    width: 80,
-                    height: 80
-                  }}
-                />
-              </IconButton>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                open={useropen}
-                onClose={handleClose}
-              >
-                <MenuItem onClick={() => linkClose('user')}>
-                  <IconButton
-                    onClick={() => linkClose('user')}
-                    color="inherit"
-                  >
-                    <AccountCircle />
-                  </IconButton>
-                  ユーザー設定
-                </MenuItem>
-                <MenuItem onClick={logoutClose}>
-                  <IconButton
-                    onClick={() => linkClose('user')}
-                    color="inherit"
-                  >
-                    <ExitToAppIcon />
-                  </IconButton>
-                  ログアウト
-                </MenuItem>
-              </Menu>
-            </React.Fragment>
-          )}
+          {signIn
+            ? <AccountIcon />
+            :
+            <div style={{ marginLeft: 'auto' }}>
+              <Button onClick={link('/user/signin')} style={{ margin: 10 }}>ログイン</Button>
+              <Button onClick={link('/user/signup')} variant='outlined' style={{ margin: 10 }}>アカウント作成</Button>
+            </div>
+          }
         </Toolbar>
       </AppBar>
       <Drawer
@@ -275,7 +162,7 @@ const NavBar: React.FC = () => {
               color: 'white'
             }}
           >
-            {theme.direction === 'ltr' ? <ChevronLeftIcon /> : <ChevronRightIcon />}
+            {theme.direction === 'ltr' ? <ChevronLeft /> : <ChevronRight />}
           </IconButton>
         </div>
         <Divider />
@@ -285,11 +172,11 @@ const NavBar: React.FC = () => {
               color="inherit"
               edge="start"
             >
-              <HomeIcon />
+              <Home />
             </IconButton>
             <ListItemText primary="ホーム" />
           </ListItem>
-          {signIn ? <SignedIn link={link} linkDrawerClose={linkDrawerClose} /> : <NotSignedIn link={link} linkDrawerClose={linkDrawerClose} />}
+          {signIn ? <SignedIn link={link} linkDrawerClose={linkDrawerClose} /> : <NotSignedIn linkDrawerClose={linkDrawerClose} />}
         </List>
         <Divider />
         <List>
@@ -301,7 +188,7 @@ const NavBar: React.FC = () => {
               color="inherit"
               edge="start"
             >
-              <HelpIcon />
+              <Help />
             </IconButton>
             <ListItemText primary="ヘルプ" />
           </ListItem>
